requirements.txt
Django==1.10.5
Markdown==2.6.8
Pillow==2.9.0
appdirs==1.4.2
argparse==1.2.1
bleach==1.5.0
bootstrap-admin==0.3.7.1
diff-match-patch==20121119
django-autocomplete-light==3.2.1
django-import-export==0.5.1
django-model-utils==2.6.1
django-password-reset==0.9
django-widget-tweaks==1.4.1
gunicorn==19.6.0
html5lib==0.9999999
packaging==16.8
psycopg2==2.6.2
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
tablib==0.11.4
wsgiref==0.1.2
xlwt==1.2.0
